competition3_digit_recognizer/cnn_digit_web_app/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
gradio
torch
numpy
pillow
